Adding a Malware Bazaar module for Enrichment
This module is built on top of the MalwareBazaar API (https://bazaar.abuse.ch/api/). A function is added to retrieve the MB API key, however it is not needed for the current features. The API key is used to send samples or send a comment while the current MSTICpy MBlookup module is used only to retrieve information about samples and more.