Resourcegraph - Incomplete list returned
Perhaps not the best solution, but it works - running the ResourceGraph driver for "PolicyResources" without the addition returns incomplete list of results, with no obvious reason. 

Noticed in the portal that not all results for a "PolicyResources"-query contain subscriptionId value, and the missing results correspond to those with "blank" values in the column.

By adding a blank string to the list, the expected results were returned.